refatorando api sector
Realocamos o endpoint de associação de setor a uma etapa para o diretório processTypeApi. Além disso, movemos as funções de callback de sectorApi para uma nova classe Sector. Esse ajuste é parte do processo de transição, onde o CPT Sector foi removido e convertido em uma classe dedicada para melhor organização e manutenção do código.
